+++ src/controllers/products.controller.js
@@ -22,6 +22,10 @@ const getOneProduct = async (req, res) => {
 
 const createProduct =  async (req, res) => {
     
+    const { role } = req.userInfo;
+
+    if (role !== 'admin') return res.status(400).json({message: 'You are not an admin.'});
+
     try {
         const { id, description, uom, stock, lastPrice, suggestedPrice } = req.body;
         
@@ -36,6 +40,9 @@ const createProduct =  async (req, res) => {
 const updateProduct = async (req, res) => {
     
     const { id } = req.params;
+    const { role } = req.userInfo;
+
+    if (role !== 'admin') return res.status(400).json({message: 'You are not an admin.'});
 
     try {
         const product = await Product.findByPk(id);
@@ -51,7 +58,11 @@ const updateProduct = async (req, res) => {
 }
 
 const deleteProduct = async (req, res) => {
+    
     const { id } = req.params;
+    const { role } = req.userInfo;
+
+    if (role !== 'admin') return res.status(400).json({message: 'You are not an admin.'});
         
     try {
         const product = await Product.destroy({
