+++ src/controllers/users.controller.js
@@ -1,6 +1,10 @@
 const User = require('../models/user.model')
 
 const getAllUsers = async (req, res) => {
+
+    const { role } = req.userInfo;
+    if (role !== 'admin') return res.status(400).json({message: 'You are not an admin.'});
+
     try {
         const users = await User.findAll();
         res.json(users)    
@@ -11,9 +15,13 @@ const getAllUsers = async (req, res) => {
 
 const getOneUser = async (req, res) => {
     
+    const { sub, role } = req.userInfo;
     const { id } = req.params;
+    
+    if (role !== 'admin' && id !== sub) return res.status(400).json({message: 'You are not an admin.'});
+        
     try {
-        const user = await User.findByPk(id)
+        const user = await User.findByPk(id);
         res.json(user)
     } catch (error) {
         return res.status(500).json({message: error.message})
@@ -21,14 +29,23 @@ const getOneUser = async (req, res) => {
 }
 
 const createUser =  async (req, res) => {
+
+    const { role } = req.userInfo;
+
+    if (role !== 'admin') return res.status(400).json({message: 'You are not an admin.'});
     
     try {
         const { id, role, password, email, firstName, lastName } = req.body;
         
         const newUser = await User.create({ id, role, password, email, firstName, lastName });
-        res.json({newUser})
+        delete newUser.dataValues.password;
+        
+        res.json({newUser});
+
     } catch (error) {
+        
         return res.status(500).json({message: error.message})
+    
     }  
 
 }
@@ -36,6 +53,9 @@ const createUser =  async (req, res) => {
 const updateUser = async (req, res) => {
     
     const { id } = req.params;
+    const { sub, role } = req.userInfo;
+
+    if (role !== 'admin' && id !== sub) return res.status(400).json({message: 'You are not an admin.'});
 
     try {
         const user = await User.findByPk(id);
@@ -51,13 +71,18 @@ const updateUser = async (req, res) => {
 }
 
 const deleteUser = async (req, res) => {
+
     const { id } = req.params;
-        
+    const { sub, role } = req.userInfo;
+
+    if (role !== 'admin' && id !== sub) return res.status(400).json({message: 'You are not an admin.'});
+
     try {
         const user = await User.destroy({
             where: { id }
         });
         return res.sendStatus(204);
+    
     } catch (error) {
         return res.status(500).json({message: error.message})
     }
